+++ Example/ViewController.swift
@@ -271,7 +271,6 @@ class ViewController: UIViewController {
         
         // Example of building highlighting in 3D.
         navigationViewController.waypointStyle = .extrudedBuilding
-        navigationViewController.detailedFeedbackEnabled = true
         
         // Show second level of detail for feedback items.
         navigationViewController.detailedFeedbackEnabled = true